ReOrganizacion de APIs a Capetas Multimedia y Operaciones
