+++ Sitio/PMMX.Sitio/Sitio.csproj
@@ -236,20 +236,20 @@
     <Compile Include="Areas\Apis\Controllers\BitacoraVentanaController.cs" />
     <Compile Include="Areas\Apis\Controllers\CarrierController.cs" />
     <Compile Include="Areas\Apis\Controllers\CategoriaController.cs" />
-    <Compile Include="Areas\Apis\Controllers\ComentariosController.cs" />
-    <Compile Include="Areas\Apis\Controllers\CRRController.cs" />
+    <Compile Include="Areas\Apis\Controllers\Operaciones\ComentariosController.cs" />
+    <Compile Include="Areas\Apis\Controllers\Operaciones\CRRController.cs" />
     <Compile Include="Areas\Apis\Controllers\EventoController.cs" />
     <Compile Include="Areas\Apis\Controllers\EventoOrigenController.cs" />
     <Compile Include="Areas\Apis\Controllers\EventoResponsableController.cs" />
-    <Compile Include="Areas\Apis\Controllers\FotosController.cs" />
-    <Compile Include="Areas\Apis\Controllers\IndicadoresController.cs" />
+    <Compile Include="Areas\Apis\Controllers\Multimedia\FotosController.cs" />
+    <Compile Include="Areas\Apis\Controllers\Operaciones\IndicadoresController.cs" />
     <Compile Include="Areas\Apis\Controllers\JustDoItController.cs" />
     <Compile Include="Areas\Apis\Controllers\InSiteLAC\KPIsController.cs" />
     <Compile Include="Areas\Apis\Controllers\ListaDistribucionController.cs" />
     <Compile Include="Areas\Apis\Controllers\LocacionController.cs" />
     <Compile Include="Areas\Apis\Controllers\LoginController.cs" />
     <Compile Include="Areas\Apis\Controllers\MantenimientoController.cs" />
-    <Compile Include="Areas\Apis\Controllers\MecanicoController.cs" />
+    <Compile Include="Areas\Apis\Controllers\Operarios\MecanicoController.cs" />
     <Compile Include="Areas\Apis\Controllers\MecanicoPorBusinessUnitsController.cs" />
     <Compile Include="Areas\Apis\Controllers\Multimedia\DefaultController.cs" />
     <Compile Include="Areas\Apis\Controllers\Operaciones\DesperdiciosController.cs" />
@@ -266,7 +266,7 @@
     <Compile Include="Areas\Apis\Controllers\SubCategoriaController.cs" />
     <Compile Include="Areas\Apis\Controllers\VentanaController.cs" />
     <Compile Include="Areas\Apis\Controllers\VQIController.cs" />
-    <Compile Include="Areas\Apis\Controllers\NoConformidadesController.cs" />
+    <Compile Include="Areas\Apis\Controllers\Operaciones\NoConformidadesController.cs" />
     <Compile Include="Areas\Apis\Controllers\EstatusController.cs" />
     <Compile Include="Areas\Apis\Controllers\StatusVentanaController.cs" />
     <Compile Include="Areas\Cuestionario\Controllers\WebApiConfig.cs" />
@@ -275,13 +275,13 @@
     <Compile Include="Areas\Cuestionario\Controllers\ModuloWorkCenterRespuestaController.cs" />
     <Compile Include="Areas\Cuestionario\Controllers\PreguntaTurnoController.cs" />
     <Compile Include="Areas\Apis\ApisAreaRegistration.cs" />
-    <Compile Include="Areas\Apis\Controllers\ActividadEnDefectoController.cs" />
-    <Compile Include="Areas\Apis\Controllers\ActividadEnParoController.cs" />
-    <Compile Include="Areas\Apis\Controllers\BussinesUnitController.cs" />
-    <Compile Include="Areas\Apis\Controllers\DefectoController.cs" />
-    <Compile Include="Areas\Apis\Controllers\EntornoController.cs" />
+    <Compile Include="Areas\Apis\Controllers\Operaciones\ActividadEnDefectoController.cs" />
+    <Compile Include="Areas\Apis\Controllers\Operaciones\ActividadEnParoController.cs" />
+    <Compile Include="Areas\Apis\Controllers\Operaciones\BussinesUnitController.cs" />
+    <Compile Include="Areas\Apis\Controllers\Operaciones\DefectoController.cs" />
+    <Compile Include="Areas\Apis\Controllers\Seguridad\EntornoController.cs" />
     <Compile Include="Areas\Apis\Controllers\GrupoPreguntasController.cs" />
-    <Compile Include="Areas\Apis\Controllers\ModuloController.cs" />
+    <Compile Include="Areas\Apis\Controllers\Operaciones\ModuloController.cs" />
     <Compile Include="Areas\Apis\Controllers\OrigenController.cs" />
     <Compile Include="Areas\Apis\Controllers\ParoController.cs" />
     <Compile Include="Areas\Apis\Controllers\PersonaController.cs" />
